% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{analyze}
\alias{analyze}
\title{Internal function to analyze one a `mp_data` based on model}
\usage{
analyze(data, alpha = 0.05, ...)
}
\description{
Internal function to analyze one a `mp_data` based on model
}
