% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{as.data.frame.mp_action}
\alias{as.data.frame.mp_action}
\title{Converts a `mp_action` class into a single row `data.frame`}
\usage{
\method{as.data.frame}{mp_action}(x, ...)
}
\description{
Converts a `mp_action` class into a single row `data.frame`
}
