% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{as.data.frame.mp_variable}
\alias{as.data.frame.mp_variable}
\title{Converts a `mp_variable` class into a single row `data.frame`}
\usage{
\method{as.data.frame}{mp_variable}(x)
}
\description{
Converts a `mp_variable` class into a single row `data.frame`
}
