% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{as.list.mp_parameters}
\alias{as.list.mp_parameters}
\title{Convert \code{\link{mp_parameters}} to a \code{\link{list}}}
\usage{
\method{as.list}{mp_parameters}(x, ...)
}
\arguments{
\item{x}{the \code{\link{mp_parameters}} to be coered.}

\item{...}{additional arguments passed to \code{\link{as.list}}}
}
\value{
a \code{\link{list}}
}
\description{
A wrapper to coerce \code{\link{mp_parameters}} to a \code{\link{list}}.
}
\examples{
# Specify model
model <- (
    outcome('Y')
    + within_predictor('X')
    + effect_size(
        icc = c(0.1, 0.2),
        within = 0.3
    )
)
# Obtain parameters and convert to a list
model |> summary() |> as.list() -> param_list
}
