% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{between_binary_predictor}
\alias{between_binary_predictor}
\title{Create Between binary predictor}
\usage{
between_binary_predictor(name, weight = 1, proportion = 0.5)
}
\description{
Create Between binary predictor
}
