% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlations}
\alias{correlations}
\title{Set up correlation defaults}
\usage{
correlations(
  within_cor = random(0.1, 0.3),
  between_cor = random(0.1, 0.3),
  randeff_cor = random(0.1, 0.3)
)
}
\description{
Set up correlation defaults
}
