% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{mp_parameters}
\alias{mp_parameters}
\title{\code{mp_parameters} Object for \code{mlmpower}}
\description{
An S3 class that contains an \code{\link[base:environment]{base::environment}} with the following objects:
\itemize{
\item \code{r2}: The population proportion of variance explained
\item \code{phi_b}: The population predictor covariance matrix for between
\item \code{phi_p}: The population within cluster covariance matrix for products
\item \code{phi_w}: The population predictor covariance matrix for within
\item \code{mean_Z}: The population mean for level-2 predictors
\item \code{mean_X}: The population mean for level-1 predictors
\item \code{var_e}: The population within residual variance
\item \code{tau}: The population level-2 covariance matrix
\item \code{gammas}: The regression coefficients
\item \code{mean_Y}: The population mean of the outcome
}
}
\seealso{
\link{parameters}
}
