% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{parameters}
\alias{parameters}
\title{Obtain \code{\link{mp_parameters}} from \code{\link{mp_data}}}
\usage{
parameters(data)
}
\arguments{
\item{data}{a \code{\link{mp_data}}.}
}
\value{
A \code{\link{mp_parameters}} object
}
\description{
Obtains the list of population parameters from a generated data set
}
\examples{
# Create Model
model <- (
    outcome('Y')
    + within_predictor('X')
    + effect_size(icc = 0.1)
)
# Set seed
set.seed(198723)
# Create data set and obtain population parameters
model |> generate(5, 50) |> parameters()
}
