% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{+.mp_parameters}
\alias{+.mp_parameters}
\title{Internal function to keep online average for `mp_parameters`}
\usage{
\method{+}{mp_parameters}(e1, e2)
}
\description{
Internal function to keep online average for `mp_parameters`
}
