% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{power_analysis}
\alias{power_analysis}
\title{Conduct a power analysis based on `mp_model`}
\usage{
power_analysis(model, replications, n_within, n_between)
}
\description{
Conduct a power analysis based on `mp_model`
}
