% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{print.mp_effect}
\alias{print.mp_effect}
\title{Prints a \code{\link{mp_effect}}}
\usage{
\method{print}{mp_effect}(x, ...)
}
\arguments{
\item{x}{a \code{\link{mp_effect}}.}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the original variable.
}
\description{
Prints a \code{\link{mp_effect}} in a human readable format.
}
