% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.mp_model}
\alias{print.mp_model}
\title{Prints a \code{\link{mp_model}}}
\usage{
\method{print}{mp_model}(x, ...)
}
\arguments{
\item{x}{a \code{\link{mp_model}}.}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the original variable.
}
\description{
Prints a \code{\link{mp_variable}} in a human readable format.
}
\examples{
print(
    outcome('Y')
    + within_predictor('X')
    + effect_size(icc = cross_sectional)
)
}
