% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate.mp_model}
\alias{simulate.mp_model}
\title{Generate a data set based on `mp_model`}
\usage{
\method{simulate}{mp_model}(object, n_within, n_between, nsim = 1)
}
\description{
Generate a data set based on `mp_model`
}
