% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.mp_model}
\alias{summary.mp_model}
\title{Provides parameter summary of a `mp_model` object
TODO improvce printing of parameters summary
HANDLE random values}
\usage{
\method{summary}{mp_model}(object, ...)
}
\description{
Provides parameter summary of a `mp_model` object
TODO improvce printing of parameters summary
HANDLE random values
}
