% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{power_analysis}
\alias{power_analysis}
\alias{mp_power}
\title{Conduct a Power Analysis Based on \code{\link{mp_model}}}
\usage{
power_analysis(model, replications, n_within, n_between, ...)
}
\arguments{
\item{model}{a \code{\link{mp_model}}.}

\item{replications}{a single positive integer of the number of replications per condition.}

\item{n_within}{an integer vector of the desired within cluster observations.}

\item{n_between}{an integer vector of the desired between cluster observations.}

\item{...}{other arguments passed to \code{\link[=analyze]{analyze()}}.}
}
\value{
A \code{mp_power} object that contains the results.
See \code{\link{print.mp_power}} for more information.
The object has the following slots:
\itemize{
\item \code{sim}: The information about the simulation
\item \code{power}: The power power results per condition.
\item \code{estimates}: The simulation summaries of the parameter estimates per condition.
\item \code{mean_parameters}: The average population parameter per condition.
}
}
\description{
This function will construct a multilevel power analysis via a Monte Carlo Simulation
based on a constructed \code{\link{mp_model}}.
}
\details{
Specifying multiple \code{n_within} and \code{n_between} will produce a full factorial simulation design.

Use a \code{mechanism} argument to pass down to the internal function that analyzes
each data set. This expects a function with the \code{\link{mp_data}} as the input,
and the function should return the modified \code{\link{mp_data}}. Be careful when
using this because it allows you to modify the population parameters, which will
be incorrect. You should only induce missing data values on variables. Missing data
on the predictors will cause listwise deletion to be used, but missing data on the
outcome will be appropriate for MAR-based mechanisms. See examples below for an
example that generates MCAR data on the outcome. See \code{\link{parameters}}
to obtain the population parameters from each data set. See \code{\link{mechanisms}}
for predefined missing data mechanisms for the outcome and examples using them.
}
\examples{
# Create Model
model <- (
    outcome('Y')
    + within_predictor('X')
    + effect_size(icc = 0.1)
)
# Set seed
set.seed(19723)
# Create data set and analyze
# Note: Generally Use more than 50 replications
model |> power_analysis(50, 5, 50)

# Induce Missing data on outcome
set.seed(19723)
model |> power_analysis(50, 5, 50, mechanism = \(data) {
    # `data` will be the `mp_data` used
    within(data, {
        # MCAR Process
        Y <- ifelse(runif(NROW(data)) < 0.5, Y, NA)
    })
}) -> powersim_mcar
}
